Move system namespace defaulting
