+++ test/defaultsystem/system.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package test
+package defaultsystem
 
 import (
 	"os"
